make separate errors.rs file and address other review comments
