+++ Cargo.toml
@@ -21,12 +21,6 @@ termios = "0.3"
 unicode-truncate = "0.1.1"
 unicode-width = "0.1.8"
 vte = "0.8.0"
-futures = "0.3.5"
-serde = { version = "1.0", features = ["derive"] }
-bincode = "1.3.1"
-structopt = "0.3"
-serde_yaml = "0.8"
-serde_json = "1.0"
 
 [dependencies.async-std]
 version = "1.3.0"
